refactor: confeti page 컴포넌트 폴더링 및 코드 수정
