+++ apps/client/src/pages/confeti/components/artist/artist-card.css.ts
