+++ apps/client/src/pages/confeti/components/artist/artist-grid.css.ts
