+++ apps/client/src/pages/confeti/components/artist/artist-section.css.ts
