+++ apps/client/src/pages/confeti/components/artist/artist-section.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import { ConcertArtist } from '../types/concert';
-import { FestivalDate } from '../types/festival';
-import ArtistGrid from './artist-grid';
-import ExpandedSection from './expanded-section';
-import * as styles from './artist-section.css';
+import { ConcertArtist } from '../../types/concert.ts';
+import { FestivalDate } from '../../types/festival.ts';
+import ArtistGrid from '@pages/confeti/components/artist/artist-grid.tsx';
+import ExpandedSection from '@pages/confeti/components/expanded/expanded-section.tsx';
+import * as styles from './artist-section.css.ts';
 
 const CONCERT_DEFAULT_ID = -1;
 const MAX_VISIBLE_ARTISTS = 4;
