+++ apps/client/src/pages/confeti/components/artist/artist-title.css.ts
