+++ apps/client/src/pages/confeti/components/expanded/expanded-section.css.ts
