+++ apps/client/src/pages/confeti/components/info/info.css.ts
