+++ apps/client/src/pages/confeti/components/performance/performance-detail.css.ts
