+++ apps/client/src/pages/confeti/components/poster/poster.css.ts
