+++ apps/client/src/pages/confeti/page/festival-detail.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
-import Poster from '../components/poster';
-import Summary from '../components/summary';
-import Info from '../components/info';
-import MoreButton from '../components/more-button';
-import PerformanceDetail from '../components/performance-detail';
-import ArtistTitle from '../components/artist-title';
-import ArtistSection from '../components/artist-section';
+import Poster from '@pages/confeti/components/poster/poster.tsx';
+import Summary from '@pages/confeti/components/summary/summary.tsx';
+import Info from '@pages/confeti/components/info/info.tsx';
+import MoreButton from '@pages/confeti/components/button/more-button.tsx';
+import PerformanceDetail from '@pages/confeti/components/performance/performance-detail.tsx';
+import ArtistTitle from '@pages/confeti/components/artist/artist-title.tsx';
+import ArtistSection from '@pages/confeti/components/artist/artist-section.tsx';
 import { FloatingButton, Footer, Spacing } from '@confeti/design-system';
 import { FESTIVAL_DETAIL } from '../mocks/confeti-detail';
 
@@ -47,7 +47,11 @@ const FestivalDetailPage = () => {
         price={festival.price}
       />
       <Spacing />
-      <PerformanceDetail isExpanded={isExpanded} />
+      <PerformanceDetail
+        isExpanded={isExpanded}
+        infoImgUrl={festival.infoImgUrl}
+        title={festival.title}
+      />
       <MoreButton
         hasShadow={true}
         isExpanded={isExpanded}
